clean up asset names, clean up code
im considering the methodology. should all assets be scaled down at the beginning, when they are loaded into the system, or should they be scaled down as necessary for scenes? it may make more sense to scale them down once to prevent needing to waste time recalculating, and it would make a lot of sense from the dev standpoint when they get accurate width/height values for a scaled surface